Use ncat instead of nc for healthcheck
Using timeout in conjunction with nc, the exit code would always be 124, no matter if the 'getinfo' command completed successfully or not. Hence I swapped nc for ncat, which is the version of netcat from the Nmap Project, which exits immediately as soon it's connected to the desired IP and port. Hence no timeout parameter for ncat is needed anymore.

Exit code is 0 if the getinfo command completed successfully, i.e. container is healthy and 1 if the getinfo command did not complete successfully within the Docker HEALTHCHECK timeout value of 5 seconds, i.e. container is not healthy